requirements.txt
python-dotenv
docx2pdf
jinja2
typer[all]
rich
InquirerPy
pydantic
pyyaml